Local flask run rework
