+++ Daratos_API/flask/.dockerignore
@@ -1,0 +1,2 @@
+__pycache__/
+env/
