+++ Daratos_API/flask/requirements.txt
@@ -1,6 +1,6 @@
 Flask==1.0.3
 nltk==3.4.5
 numpy==1.16.2
-tensorflow==2.0.0
+tensorflow==1.14.0
 firebase-admin==3.1.0
-uWSGI==2.0.10
+uWSGI==2.0.18
