+++ Daratos_API/nginx/dockerfile
@@ -1,0 +1,8 @@
+# Use the Nginx image
+FROM nginx
+
+# Remove the default nginx.conf
+RUN rm /etc/nginx/conf.d/default.conf
+
+# Replace with our own nginx.conf
+COPY nginx.conf /etc/nginx/conf.d/
