+++ Daratos_API/nginx/nginx.conf
@@ -1,0 +1,10 @@
+server {
+
+    listen 80;
+
+    location / {
+        include uwsgi_params;
+        uwsgi_pass flask:8080;
+    }
+
+}
